finding race condition in server for win condition

+++ server/listeners.js
@@ -11,15 +11,25 @@ const listeners = function(io, socket){
 
   socket.on('disconnect', function(){
     console.log('socket id ' + socket.id + ' has disconnected.');
-    store.dispatch(removePlayer(socket.id));
+    console.log('gmExist:', store.getState().gmExist)
+    if (store.getState().gmExist){
+      store.dispatch(removePlayer(socket.id));
+    }
   });
 
   socket.on('send_all_data', (data) => {
-    console.log(store.getState())
+    // console.log(store.getState())
     if (data.gameMode === 'survivor') {
       store.dispatch(receiveClientData(socket.id, data));
     } else {
       store.dispatch(updateMonsters(data.monsters));
+      if (store.getState().gmExist) {
+        if (data.gameMode === 'survivor') {
+          store.dispatch(receiveClientData(socket.id, data));
+        } else {
+          store.dispatch(updateMonsters(data.monsters));
+        }
+      }
     }
   })
 }
